Mod to select working random locations in GTAV
DeepGTAV currently chooses a random start location numerically, but if the location is not possible it defaults to a fixed location. The proposed code selects a working random location from a csv of 350 random working locations. The project will not allow me to upload the csv file or make a pull request so I'll have to send you the csv separately to add to the project. Message me on how best to do that or give me pull request/upload proposal rights and I'll do it myself. Cheers e.